Be strict with rdp_protocols parsing, allow correlation info after RDP_NEGREQ
Since the rdp_protocols field has to be identical between the negreq
and the server TSUDs, we might as well error out if there's anything
in it that we don't recognise (rather than not copy it and die
anyway in a way that's hard to see).

The correlation info fix seems to help with MSTSC in some configs.
Haven't seen any other client send it yet.